# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Patient Research Data Collection API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Object content.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/ensure-bucket:
    post:
      tags: [Storage]
      operationId: ensureBucket
      summary: Ensure storage bucket exists
      responses:
        '200':
          description: Bucket is ready.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '404':
          description: Bucket not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/images/import:
    post:
      tags: [Storage]
      operationId: importImageFromUrl
      summary: Import an image from a URL into storage
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportImageFromUrlRequest'
      responses:
        '200':
          description: Image imported successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportImageFromUrlResponse'
        '400':
          description: Invalid request or URL does not point to an image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/images/by-patient:
    post:
      tags: [Storage]
      operationId: importImageByPatient
      summary: Import an image from a URL and associate with a patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ImportImageByPatientRequest'
      responses:
        '200':
          description: Image imported successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ImportImageByPatientResponse'
        '400':
          description: Invalid request or URL does not point to an image.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/images/by-patient/{patientId}:
    get:
      tags: [Storage]
      operationId: getPatientImages
      summary: List all images associated with a patient ID
      parameters:
        - name: patientId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Images found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientImagesResponse'
        '500':
          description: Server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /storage/images/search:
    post:
      tags: [Storage]
      operationId: searchImages
      summary: Search for images by identifier or filename
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                identifier:
                  type: string
                  description: Image identifier to search (e.g., patient ID, study ID)
                patientId:
                  type: string
                  description: Existing patient ID to attach images to
                filename:
                  type: string
                  description: Filename pattern to search for
              additionalProperties: false
      responses:
        '200':
          description: Images found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchImagesResponse'
        '404':
          description: No images found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  schemas:
...
    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
        size:
          type: integer
          minimum: 1
        contentType:
          type: string
          minLength: 1
    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
        objectPath:
          type: string
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ImportImageFromUrlRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        filename:
          type: string
    ImportImageFromUrlResponse:
      type: object
      required: [objectPath, contentType]
      properties:
        objectPath:
          type: string
        contentType:
          type: string
    ImportImageByPatientRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
        filename:
          type: string
        patientId:
          type: string
    ImportImageByPatientResponse:
      type: object
      required: [objectPath, patientImages]
      properties:
        objectPath:
          type: string
        patientImages:
          type: array
          items:
            type: string
    PatientImagesResponse:
      type: object
      required: [patientId, images]
      properties:
        patientId:
          type: string
        images:
          type: array
          items:
            type: string
    SearchImagesResponse:
      type: object
      required: [objectPath]
      properties:
        objectPath:
          type: string
        patientImages:
          type: array
          items:
            type: string
        attachmentPatientId:
          type: string
...